requirements.txt
flet>=0.10.0
psycopg[binary]>=3.1
psycopg_pool>=3.1
python-dotenv>=1.0
